Ajout de la fonctionnalité de paiement et d'ajout des commandes et ar…

+++ src/controleur/articleControleur.java
@@ -30,7 +30,7 @@ public class articleControleur implements ActionListener {
                 return;
             }
 
-            double prixTotal = article.calculerPrix(quantite);
+            float prixTotal = article.calculerPrix(quantite);
 
             Panier panier = Panier.getInstance();
             panier.ajouterArticle(article, quantite, prixTotal);
